refactor(storefront): move getStoreSlugFromHost
