+++ apps/storefront/src/app/[lang]/layout.tsx
@@ -20,8 +20,8 @@ import { StoreSelector } from "@/components/layout/store-selector";
 import { StoreSelectorBubble } from "@/components/layout/store-selector-bubble";
 import { STORE_HEADER_HEIGHT_PX } from "@/lib/constants";
 import { handleAPIError } from "@/lib/error";
-import { getStoreSlug } from "@/lib/get-store-slug";
 import { getQueryClient, orpc } from "@/lib/orpc";
+import { getStoreSlug } from "@/shared/lib/store/slug-retrieval.util";
 
 const inter = Inter({
   variable: "--font-sans-latin",
